logging: refactor: move output handling to a separate file for CN
