+++ components/logging/output.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <mavlink/mavlink_types.h>
+
+void handle_message(const mavlink_message_t* msg);
